# api_context/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Market Sentiment Intelligence API",
    "description": "Near real-time sentiment analysis platform for traders and crypto users.\n\n## Quick Start\n1. Call `POST /api/auth/signup` to create an account\n2. Call `POST /api/auth/login` to get a JWT token\n3. Use the token in the `Authorization: Bearer <token>` header for all protected routes\n\n## Pre-filled Values\nThis spec contains temporary example values so you can test immediately in Postman without changing anything.",
    "version": "1.0.0",
    "contact": {
      "name": "Market Sentiment Intelligence"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Get your token from POST /api/auth/login"
      }
    },
    "schemas": {
...
      "Strategy": {
        "type": "object",
        "properties": {
          "id":          { "type": "string", "format": "uuid" },
          "userId":      { "type": "string", "format": "uuid" },
          "name":        { "type": "string", "example": "My Strategy" },
          "description": { "type": "string", "nullable": true },
          "isActive":    { "type": "boolean", "example": true },
          "config":      { "$ref": "#/components/schemas/StrategyConfig" },
          "createdAt":   { "type": "string", "format": "date-time" },
          "updatedAt":   { "type": "string", "format": "date-time" }
        }
      },
      "StrategyConfig": {
        "type": "object",
        "properties": {
          "keywordsPositive":    { "type": "array", "items": { "type": "string" }, "example": ["bullish", "moon"] },
          "keywordsNegative":    { "type": "array", "items": { "type": "string" }, "example": ["crash", "dump"] },
          "impactThreshold":     { "type": "number", "example": 70 },
          "confidenceThreshold": { "type": "number", "example": 0.6 },
          "sentimentWeight":     { "type": "number", "example": 0.5 },
          "impactWeight":        { "type": "number", "example": 0.5 },
          "sentimentThreshold":  { "type": "number", "example": 0.2 },
          "categories": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SentimentCategory" },
            "example": ["SOCIAL_BUZZ", "NEWS", "RUMOR", "WHALE_ACTIVITY"]
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "integer", "example": 401 },
          "timestamp":  { "type": "string", "format": "date-time" },
          "path":       { "type": "string", "example": "/api/analysis/deep" },
          "message":    { "type": "object" }
        }
      },
...
      "SentimentCategory":{ "type": "string", "enum": ["SOCIAL_BUZZ", "NEWS", "RUMOR", "WHALE_ACTIVITY"] },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid JWT token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/strategies/strategy": {
      "get": {
        "tags": ["Strategies"],
        "summary": "Get user's active strategy config (returns defaults if none)",
        "operationId": "getActiveStrategy",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Active strategy config",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/StrategyConfig" } }
            }
          }
        }
      }
    },
    "/api/strategies/strategy/update": {
      "post": {
        "tags": ["Strategies"],
        "summary": "Create or update the active strategy (partial updates supported)",
        "operationId": "upsertActiveStrategy",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "keywordsPositive":    { "type": "array", "items": { "type": "string" }, "example": ["bullish", "moon", "breakout"] },
                  "keywordsNegative":    { "type": "array", "items": { "type": "string" }, "example": ["crash", "dump", "bearish"] },
                  "impactThreshold":     { "type": "number", "minimum": 0, "maximum": 100, "example": 65 },
                  "confidenceThreshold": { "type": "number", "minimum": 0, "maximum": 1, "example": 0.5 },
                  "sentimentWeight":     { "type": "number", "minimum": 0, "maximum": 1, "example": 0.6 },
                  "impactWeight":        { "type": "number", "minimum": 0, "maximum": 1, "example": 0.4 },
                  "sentimentThreshold":  { "type": "number", "minimum": 0, "maximum": 1, "example": 0.2 },
                  "categories": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/SentimentCategory" },
                    "example": ["SOCIAL_BUZZ", "NEWS", "WHALE_ACTIVITY"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated strategy",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Strategy" } }
            }
          }
        }
      }
    },
    "/api/strategies": {
      "get": {
        "tags": ["Strategies"],
        "summary": "List all user strategies",
        "operationId": "listStrategies",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Array of strategies",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Strategy" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Strategies"],
        "summary": "Create a new strategy",
        "operationId": "createStrategy",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "config"],
                "properties": {
                  "name":        { "type": "string", "example": "Aggressive Bull Strategy" },
                  "description": { "type": "string", "example": "High impact threshold with bullish bias" },
                  "config":      { "$ref": "#/components/schemas/StrategyConfig" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Strategy created",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Strategy" } }
            }
          }
        }
      }
    },
    "/api/strategies/{id}": {
      "get": {
        "tags": ["Strategies"],
        "summary": "Get a strategy by ID",
        "operationId": "getStrategyById",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "example": "550e8400-e29b-41d4-a716-446655440000" }
        ],
        "responses": {
          "200": {
            "description": "Strategy object",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Strategy" } }
            }
          }
        }
      },
      "put": {
        "tags": ["Strategies"],
        "summary": "Update a strategy",
        "operationId": "updateStrategy",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "example": "550e8400-e29b-41d4-a716-446655440000" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name":     { "type": "string", "example": "Updated Strategy Name" },
                  "isActive": { "type": "boolean", "example": true },
                  "config":   { "$ref": "#/components/schemas/StrategyConfig" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated strategy",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Strategy" } }
            }
          }
        }
      },
      "delete": {
        "tags": ["Strategies"],
        "summary": "Delete a strategy",
        "operationId": "deleteStrategy",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "example": "550e8400-e29b-41d4-a716-446655440000" }
        ],
        "responses": { "200": { "description": "Strategy deleted" } }
      }
    },
...
    "/api/strategies/generate": {
      "post": {
        "tags": ["StrategyGen"],
        "summary": "Generate 3 AI-powered strategies based on investor profile",
        "description": "Uses GPT-4o-mini to generate Conservative, Balanced, and Aggressive strategies. Falls back to rule-based if OpenAI not configured.",
        "operationId": "generateStrategies",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["assets"],
                "properties": {
                  "assets": { "type": "array", "items": { "type": "string" }, "example": ["RELIANCE", "TCS"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Array of 3 generated strategies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name":             { "type": "string", "example": "Conservative Guardian" },
                      "description":      { "type": "string" },
                      "riskLevel":        { "type": "string", "enum": ["LOW", "MEDIUM", "HIGH"] },
                      "estimatedWinRate": { "type": "string", "example": "72%" },
                      "rationale":        { "type": "string" },
                      "config":           { "$ref": "#/components/schemas/StrategyConfig" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  }
}